chore: moved ui state for inputs into the viewmodel
Moved UI state for the inputs into the view model.
Added methods to update the UI state based on the
inputs. Collect distinct flows from the input changes
to trigger a search. Using a courountine job to
check if there is already an active one running.
If the search ui changes while there is an active
job. It will cancel the job to execute a new request.
Updated README and unit tests.